Seperate top paid players into it's own class.
